requirements.txt
# requirements.txt
# OpenSearch Korean Neural Sparse Model Dependencies
# Python 3.12 Compatible | GPU: Tesla T4 (Compute Capability 7.5)

# Core ML frameworks (CUDA 12.1 support for Tesla T4)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Transformers and Hugging Face (Python 3.12 compatible stable versions)
transformers==4.46.3
datasets==2.21.0
accelerate==1.1.1
huggingface-hub==0.25.2
tokenizers==0.20.3
safetensors==0.4.5

# Korean NLP (optional - not required for core functionality)
# konlpy==0.6.0  # Optional: Only needed for advanced Korean tokenization
# mecab-python3==1.0.9  # ⚠️ ARM systems: Installation may fail, not required

# Data science (Python 3.12 compatible)
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Progress and utilities
tqdm==4.66.6

# Jupyter (optional)
jupyter==1.1.1
ipywidgets==8.1.5
notebook==7.2.2

# Typing extensions for Python 3.12
typing-extensions==4.12.2

# Additional utilities
sentencepiece==0.2.0
protobuf==5.28.3
sentence-transformers==3.3.1

# Phase 1: Loss functions and training improvements
# (No additional dependencies needed - using PyTorch built-ins)

# Phase 3: Hard negative mining
rank-bm25==0.2.2
FlagEmbedding==1.2.10
faiss-cpu==1.9.0.post1

# Preprocessing pipeline
click>=8.1.0
datasketch>=1.6.0

# Phase 4: Clustering for synonym discovery
scipy==1.14.1  # For clustering algorithms
hdbscan==0.8.39  # Advanced clustering (optional)

# Configuration management
pyyaml==6.0.2
mwparserfromhell==0.6.6

# Environment variable management
python-dotenv>=1.0.0

# LLM inference (ARM aarch64 compatible, Python 3.12)
# Using FP8 quantization: Transformers 4.46.3+ native support, no external deps
# AWQ quantization NOT supported: autoawq is deprecated and doesn't support ARM aarch64
# FP8 provides good compression (~30GB vs ~60GB BF16) with native ARM support
# llama-cpp-python==0.3.4  # Optional: gpt-oss-20b GGUF support (requires build)
